+++ admin/sql/update.sql
@@ -38,7 +38,7 @@ INSERT INTO `config` (`name`, `value`) VALUES
 ('new_tanks', '172800');
 
 CREATE TABLE IF NOT EXISTS `multiclan` (
-  `id` int(25) NOT NULL,
+  `id` bigint(225) NOT NULL,
   `prefix` varchar(64) NOT NULL,
   `sort` int(9) NOT NULL,
   `main` int(1) NOT NULL DEFAULT '0',
